refactor: move response context methods to new file
